Improved file structure and improved modularity
